A bit of small clean up
