+++ providers/vsphere/resources/datacenter.go
@@ -10,11 +10,11 @@ import (
 	"github.com/vmware/govmomi/object"
 	"go.mondoo.com/cnquery/v11/llx"
 	"go.mondoo.com/cnquery/v11/providers-sdk/v1/plugin"
+	"go.mondoo.com/cnquery/v11/providers-sdk/v1/util/convert"
 	"go.mondoo.com/cnquery/v11/providers/vsphere/connection"
 	"go.mondoo.com/cnquery/v11/providers/vsphere/resources/resourceclient"
 	"go.mondoo.com/cnquery/v11/types"
 )
-
 
 func newVsphereHostResources(vClient *resourceclient.Client, runtime *plugin.Runtime, vhosts []*object.HostSystem) ([]interface{}, error) {
 	mqlHosts := make([]interface{}, len(vhosts))
@@ -30,12 +30,11 @@ func newVsphereHostResources(vClient *resourceclient.Client, runtime *plugin.Run
 			return nil, err
 		}
 
-		// Extract tags from hostInfo.Tag field (simpler than vAPI REST client)
-		tags := extractTagKeys(hostInfo.Tag)
-
 		var name string
+		var tags []string
 		if hostInfo != nil {
 			name = hostInfo.Name
+			tags = extractTagKeys(hostInfo.Tag)
 		}
 
 		mqlHost, err := CreateResource(runtime, "vsphere.host", map[string]*llx.RawData{
@@ -43,7 +42,7 @@ func newVsphereHostResources(vClient *resourceclient.Client, runtime *plugin.Run
 			"name":          llx.StringData(name),
 			"properties":    llx.DictData(props),
 			"inventoryPath": llx.StringData(h.InventoryPath),
-			"tags":          llx.ArrayData(stringSliceToInterface(tags), types.String),
+			"tags":          llx.ArrayData(convert.SliceAnyToInterface(tags), types.String),
 		})
 		if err != nil {
 			return nil, err
