+++ providers/vsphere/resources/vm.go
@@ -6,31 +6,13 @@ package resources
 import (
 	"github.com/vmware/govmomi/object"
 	"github.com/vmware/govmomi/vim25/mo"
-	vmwaretypes "github.com/vmware/govmomi/vim25/types"
 	"go.mondoo.com/cnquery/v11/llx"
 	"go.mondoo.com/cnquery/v11/providers-sdk/v1/plugin"
+	"go.mondoo.com/cnquery/v11/providers-sdk/v1/util/convert"
 	"go.mondoo.com/cnquery/v11/providers/vsphere/connection"
 	"go.mondoo.com/cnquery/v11/providers/vsphere/resources/resourceclient"
 	"go.mondoo.com/cnquery/v11/types"
 )
-
-// extractTagKeys extracts tag keys from vmware Tag slice
-func extractTagKeys(tags []vmwaretypes.Tag) []string {
-	tagKeys := make([]string, len(tags))
-	for i, tag := range tags {
-		tagKeys[i] = tag.Key
-	}
-	return tagKeys
-}
-
-// stringSliceToInterface converts []string to []interface{}
-func stringSliceToInterface(slice []string) []interface{} {
-	result := make([]interface{}, len(slice))
-	for i, v := range slice {
-		result[i] = v
-	}
-	return result
-}
 
 func newMqlVm(runtime *plugin.Runtime, vm *object.VirtualMachine, vmInfo *mo.VirtualMachine) (*mqlVsphereVm, error) {
 	props, err := resourceclient.VmProperties(vmInfo)
@@ -39,19 +21,18 @@ func newMqlVm(runtime *plugin.Runtime, vm *object.VirtualMachine, vmInfo *mo.Vir
 	}
 
 	var name string
+	var tags []string
 	if vmInfo != nil && vmInfo.Config != nil {
 		name = vmInfo.Config.Name
+		tags = extractTagKeys(vmInfo.Tag)
 	}
-
-	// Extract tags from vmInfo.Tag field (simpler than vAPI REST client)
-	tags := extractTagKeys(vmInfo.Tag)
 
 	mqlVm, err := CreateResource(runtime, "vsphere.vm", map[string]*llx.RawData{
 		"moid":          llx.StringData(vm.Reference().Encode()),
 		"name":          llx.StringData(name),
 		"properties":    llx.DictData(props),
 		"inventoryPath": llx.StringData(vm.InventoryPath),
-		"tags":          llx.ArrayData(stringSliceToInterface(tags), types.String),
+		"tags":          llx.ArrayData(convert.SliceAnyToInterface(tags), types.String),
 	})
 	if err != nil {
 		return nil, err
